# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Produtos",
        "description": "Documentação da API de Produtos",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/product": {
            "get": {
                "summary": "Retorna uma lista de produtos",
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Produto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Cria um novo produto",
                "tags": [
                    "Products"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produto"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Produto ja existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Já existe esse produto",
                                                "description": "Já existe esse produto"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao criar produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Erro ao criar produto",
                                                "description": "Erro ao criar produto"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/joke": {
            "get": {
                "tags": [
                    "Joker"
                ],
                "summary": "retorna uma piada aleatoria",
                "operationId": "joke",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "joke": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Produto": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "preco": {
                        "type": "string"
                    },
                    "criadoPor": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "__v": {
                        "type": "integer"
                    },
                    "atualizadoPor": {
                        "type": "string"
                    }
                },
                "required": [
                    "nome",
                    "descricao",
                    "preco"
                ]
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}